Migrate the disjunct structure
